CircleImg component is moved to global components + added placeholder in case image is null
