--- src/components/project/CircleImg.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function CircleImg(props) {
-  return (
-    <img className='card-img-top circle-img' src={props.logo} alt='client' />
-  );
-}
-
-export default CircleImg;
