+++ src/components/project/ClientInfo.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import CircleImg from './CircleImg';
+import CircleImg from '../global/CircleImg';
 import {config} from '../../util/config.js';
 
 export default class ClientInfo extends React.Component {
